Phase 5 (1/N): split nest-format leaf modules into directories
encoding.rs (321) → encoding/{mod, zstd_codec, float16, int8}.rs
layout.rs (346) → layout/{mod, header, section_entry, footer}.rs
manifest.rs (368) → manifest/{mod, validate, canonical}.rs
sections.rs (409) → sections/{mod, codec, chunk_ids, canonical, spans, provenance, contract}.rs

Zero behavior change. Each test from the originals split among the
new files plus a few module-local tests added. Re-export from each
mod.rs preserves the public API.

Largest file in nest-format/src after this pass:
  writer.rs 626 (next)
  reader.rs 413 (after writer)
  manifest/mod.rs 240
  layout/mod.rs 164

cargo test -p nest-format: still 47 + 19 + 8 (...) green.
cargo clippy --workspace -- -D warnings: clean.
